Move homepage to nextjs file-system based page
