+++ examples/next/faust-nx/wp-templates/index.js
@@ -1,10 +1,8 @@
 import page from './page';
 import archive from './archive';
 import single from './single';
-import home from './home';
 
 export default {
-  index: home,
   page,
   archive,
   single
